Add logic to check if button pressed will be logged, reorder functions
I moved the if statements checking if the clear or backspace buttons were
pressed, from main, into a function- isButtonToLog.

I also moved functions around.

I removed updateOperationsSinceClear from the clear function as it
is no longer needed.

I also renamed showButtonPressed to processButton to better match the
intention of the function with the new order and logic.